Remove participate wallet process for share xpubs simple panel
